fix tests after dir change
